# go/model_gravity_obj.go
/*
 * PiHole Management API
 *
 * A management webservice to be run locally on a PiHole instance
 *
 * API version: 1.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi


import (
	"encoding/json"
)



type GravityObj struct {

	Id int32 `json:"id,omitempty"`

	Address string `json:"address"`

	Comment string `json:"comment,omitempty"`
}

// UnmarshalJSON sets *m to a copy of data while respecting defaults if specified.
func (m *GravityObj) UnmarshalJSON(data []byte) error {

	type Alias GravityObj // To avoid infinite recursion
    return json.Unmarshal(data, (*Alias)(m))
}

// AssertGravityObjRequired checks if the required fields are not zero-ed
func AssertGravityObjRequired(obj GravityObj) error {
	elements := map[string]interface{}{
		"address": obj.Address,
	}
	for name, el := range elements {
		if isZero := IsZeroValue(el); isZero {
			return &RequiredError{Field: name}
		}
	}

	return nil
}

// AssertGravityObjConstraints checks if the values respects the defined constraints
func AssertGravityObjConstraints(obj GravityObj) error {
	return nil
}